go lint err fix (#2)
* fix prompt and llm


* fix script


* fix prompt and llm


* fix go lint


* fix foundation dao name


* fix go lint


* fix go imports


* fix golint


* fix: golangci-lint err

* fix: golangci timeout

* prompt basic delete commit_status


* fix: pool err return

* fix: json repair pkg

---------

+++ backend/modules/llm/application/convertor/manage.go
@@ -70,8 +70,8 @@ func AbilityImageDO2DTO(a *entity.AbilityImage) *manage.AbilityImage {
 	return &manage.AbilityImage{
 		URLEnabled:    ptr.Of(a.URLEnabled),
 		BinaryEnabled: ptr.Of(a.BinaryEnabled),
-		MaxImageSize:  ptr.Of(int64(a.MaxImageSize)),
-		MaxImageCount: ptr.Of(int64(a.MaxImageCount)),
+		MaxImageSize:  ptr.Of(a.MaxImageSize),
+		MaxImageCount: ptr.Of(a.MaxImageCount),
 	}
 }
